Move postprocess logic to a mixin
