+++ api/preprint_providers/views.py
@@ -9,7 +9,7 @@ from framework.auth.oauth_scopes import CoreScopes
 from website.models import Node, Subject, PreprintService, PreprintProvider
 
 from api.base import permissions as base_permissions
-from api.base.filters import DjangoFilterMixin, ODMFilterMixin
+from api.base.filters import PreprintFilterMixin, ODMFilterMixin
 from api.base.views import JSONAPIBaseView
 from api.base.pagination import MaxSizePagination
 from api.base.utils import get_object_or_error, get_user_auth
@@ -142,7 +142,7 @@ class PreprintProviderDetail(JSONAPIBaseView, generics.RetrieveAPIView):
         return get_object_or_error(PreprintProvider, self.kwargs['provider_id'], display_name='PreprintProvider')
 
 
-class PreprintProviderPreprintList(JSONAPIBaseView, generics.ListAPIView, DjangoFilterMixin):
+class PreprintProviderPreprintList(JSONAPIBaseView, generics.ListAPIView, PreprintFilterMixin):
     """Preprints from a given preprint_provider. *Read Only*
 
     To update preprints with a given preprint_provider, see the `<node_id>/relationships/preprint_provider` endpoint
@@ -198,20 +198,6 @@ class PreprintProviderPreprintList(JSONAPIBaseView, generics.ListAPIView, Django
 
     view_category = 'preprint_providers'
     view_name = 'preprints-list'
-
-    def postprocess_query_param(self, key, field_name, operation):
-        if field_name == 'provider':
-            operation['source_field_name'] = 'provider___id'
-
-        if field_name == 'id':
-            operation['source_field_name'] = 'guids___id'
-
-        if field_name == 'subjects':
-            try:
-                Subject.objects.get(_id=operation['value'])
-                operation['source_field_name'] = 'subjects___id'
-            except Subject.DoesNotExist:
-                operation['source_field_name'] = 'subjects__text'
 
     # overrides DjangoFilterMixin
     def get_default_django_query(self):
